add slider for rewiews
