+++ src/main.js
@@ -1,3 +1,4 @@
 import './js/menu.js'
 import './js/modal-shopnow.js'
 import './js/modal-shopnow-mobile.js'
+import './js/slider-rewiews.js';
