+++ src/partials/reviews.html
@@ -60,7 +60,6 @@
         </li>
       </ul>
       <button class="slider-button next-button">&#10095;</button>
-      <!-- <div class="slider-indicators"></div> -->
     </div>
     
   </div>
